requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
selenium>=4.15.2
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
requests>=2.31.0
webdriver-manager>=4.0.1
pydantic-settings>=2.0.3 
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2